Refactor Favorites component and useItemStoreContext hook
